advanced_ai_agents/multi_agent_apps/ai_negotiation_battle_simulator/backend/requirements.txt
google-adk>=1.5.0
google-genai>=1.0.0
ag-ui-adk>=0.1.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
pydantic>=2.0.0
